condense code and remove hardcoded childvalue from ui
