+++ Localize/lang/strings.json
@@ -228,6 +228,7 @@
   "9JuvQ4": "Returns a data URI of a value",
   "9Jv3+1": "Returns an object with an additional property value pair",
   "9SuA/E": "12",
+  "9YGcT2": "ChildErrors",
   "9atGYe": "Add",
   "9djnqI": "Returns the result from adding the two numbers",
   "9pdXmX": "(UTC+01:00) Amsterdam, Berlin, Bern, Rome, Stockholm, Vienna",
@@ -920,6 +921,7 @@
   "_9JuvQ4.comment": "Label for description of custom dataUri Function",
   "_9Jv3+1.comment": "Label for description of custom addProperty Function",
   "_9SuA/E.comment": "Hour of the day",
+  "_9YGcT2.comment": "The title of the child errors field in the static result parseJson action",
   "_9atGYe.comment": "Insert Expression",
   "_9djnqI.comment": "Label for description of custom add Function",
   "_9pdXmX.comment": "Timezone value ",
