+++ libs/services/intl/src/compiled-lang/strings.en-XA.json
@@ -1495,6 +1495,12 @@
       "value": "12"
     }
   ],
+  "9YGcT2": [
+    {
+      "type": 0,
+      "value": "ḈḫíĺḋḔŕŕŏŕś"
+    }
+  ],
   "9atGYe": [
     {
       "type": 0,
